requirements.txt
aiofiles
fastapi
gunicorn
python-dotenv
uvicorn
websockets